It's the classic FizzBuzz Test
"Write a program that prints the numbers from 1 to 100. But for
multiples of three print “Fizz” instead of the number and for the
multiples of five print “Buzz”. For numbers which are multiples of both
three and five print “FizzBuzz”."

Analysis is included to show understanding of how organization of code
affects efficiency.